+++ Assets/Ludo Masters/Scripts/UI/ludoBet.cs
@@ -74,7 +74,7 @@ public class ludoBet : MonoBehaviour
     public GameConfigrationController controller;
     private void OnPlayClick()
     {
-        float totalPoints = TotalPoints.instance.gamePoints;
+        int totalPoints = TotalPoints.instance.gamePoints;
 
         // Check if _count is greater than 0 and less than or equal to totalPoints
         if (_count > 0 && _count <= totalPoints)
